refactoring urls, fixing crud
