+++ requirements.txt
@@ -2,3 +2,4 @@ Flask==2.2.2
 Flask-SQLAlchemy==2.5.1
 Flask-WTF==1.0.1
 WTForms-Alchemy==0.18.0
+Flask-Migrate==3.1.0
