+++ webapp/templates/order_attachment_list.html
@@ -4,7 +4,7 @@
            
     <div class="row">
         <div class="col-12">
-            {% if form.id.data > 0 %}
+            {% if form.id.data %}
             <a href="{{ url_for('new_order_attachment', order_id=form.id.data) }}">new attachment</a>
             {% else %}
             <p>new attachment</p>
