+++ webapp/templates/point_item.html
@@ -16,12 +16,9 @@ web-tracoma/point
     <form action="{{ url_for('save_point') }}" method="post">
         {{ form.csrf_token }}
         <div class="row">
-
-            <div class="input-group input-group-sm mb-3 visually-hidden">
-                <span class="input-group-text" id="inputGroup-sizing-sm">{{ form.id.label() }}</span>
-                <input type="text" class="form-control" aria-label="Sizing example input"
-                    aria-describedby="inputGroup-sizing-sm" name="id" value="{{ form.id.data }}">
-            </div>
+            
+            {% include 'including_id.html' %}
+            
             <div class="row">
                 <div class="col-10">
                     <div class="input-group input-group-sm mb-3">
