+++ webapp/templates/receipt_attachment_item.html
@@ -13,11 +13,8 @@ web-tracoma/receipt attachment
         {{ form.csrf_token }}
         <div class="row">
 
-            <div class="input-group input-group-sm mb-3 visually-hidden">
-                <span class="input-group-text" id="inputGroup-sizing-sm">{{ form.id.label() }}</span>
-                <input type="text" class="form-control" aria-label="Sizing example input"
-                    aria-describedby="inputGroup-sizing-sm" name="id" value="{{ form.id.data }}">
-            </div>
+            {% include 'including_id.html' %}
+            
             <div class="input-group input-group-sm mb-3 visually-hidden">
                 <span class="input-group-text" id="inputGroup-sizing-sm">{{ form.receipt_id.label() }}</span>
                 <input type="text" class="form-control" aria-label="Sizing example input"
