+++ webapp/templates/receipt_item.html
@@ -17,11 +17,7 @@ web-tracoma/receipt
         {{ form.csrf_token }}
         <div class="row">
 
-            <div class="input-group input-group-sm mb-3 visually-hidden">
-                <span class="input-group-text" id="inputGroup-sizing-sm">{{ form.id.label() }}</span>
-                <input type="text" class="form-control" aria-label="Sizing example input"
-                    aria-describedby="inputGroup-sizing-sm" name="id" value="{{ form.id.data }}">
-            </div>
+            {% include 'including_id.html' %}
 
             <div class="input-group input-group-sm mb-3">
                 <span class="input-group-text" id="inputGroup-sizing-sm">{{ form.storage_id.label() }}</span>
