+++ webapp/urls/__init__.py
@@ -1,0 +1,7 @@
+from webapp.urls.common_urls import add_common_urls
+from webapp.urls.order_urls import add_order_urls
+from webapp.urls.receipt_urls import add_receipt_urls
+from webapp.urls.movement_urls import add_movement_urls
+from webapp.urls.issuance_urls import add_issuance_urls
+from webapp.urls.storage_urls import add_storage_urls
+from webapp.urls.point_urls import add_point_urls
