+++ webapp/views/cargo_movement/forms.py
@@ -13,7 +13,7 @@ class MovementForm(ModelForm):
     class Meta:
         model = CargoMovement
 
-    id = IntegerField(label="Id: ", validators=[DataRequired()], name="id", default=-1)
+    id = IntegerField(label="Id: ", name="id")
     title = StringField(label="Title: ", name="title", default="")
     shipper_id = IntegerField(label="Shipper: ", validators=[DataRequired()], name="shipper_id")
     consignee_id = IntegerField(label="Consignee: ", validators=[DataRequired()], name="consignee_id")
@@ -26,7 +26,7 @@ class AttachmentForm(ModelForm):
     class Meta:
         model = AttachmentCargoMovement
 
-    id = IntegerField(label="Id: ", validators=[DataRequired()], name="id", default=-1)
+    id = IntegerField(label="Id: ", name="id")
     movement_id = IntegerField(label="Movement_id: ", validators=[DataRequired()], name="movement_id")
     order_id = IntegerField(label="Order: ", validators=[DataRequired()], name="order")
