+++ webapp/views/issuance_of_cargo/forms.py
@@ -13,7 +13,7 @@ class IssuanceForm(ModelForm):
     class Meta:
         model = IssuanceOfCargo
 
-    id = IntegerField(label="Id: ", validators=[DataRequired()], name="id", default=-1)
+    id = IntegerField(label="Id: ", name="id")
     title = StringField(label="Title: ", name="title", default="")
     storage_id = IntegerField(label="Storage: ", validators=[DataRequired()], name="storage")
     cost = FloatField(label="Cost: ", name="cost", default=.0)
@@ -25,7 +25,7 @@ class AttachmentForm(ModelForm):
     class Meta:
         model = AttachmentIssuanceOfCargo
 
-    id = IntegerField(label="Id: ", validators=[DataRequired()], name="id", default=-1)
+    id = IntegerField(label="Id: ", name="id")
     issuance_id = IntegerField(label="Issuance_id: ", validators=[DataRequired()], name="issuance_id")
     order_id = IntegerField(label="Order: ", validators=[DataRequired()], name="order")
